backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
langchain-core
langchain-openai
langchain-community
langgraph
typing_extensions
python-multipart
azure-storage-blob
azure-search-documents
azure-core
unstructured[all-docs]
azure-identity
python-magic
azure-data-tables
transformers
torch
sentence-transformers
openai